Add SimplePackageCacheManager
